Avoid pipenv completion errors with a non-default shell
Pipenv uses `SHELL` to determine the language output by `pipenv --completion`. If using fish as a non-default shell, `SHELL` will still be set to (e.g.) `/bin/bash`, resulting in `pipenv --completion` outputting bash script, which gives an error in fish.

As described [here](https://github.com/pypa/pipenv/issues/1762), the pipenv maintainers require pipenv to be used with a default shell, but this use of env is viewed as a suitable workaround.